# docs/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeChat Channels Download API",
    "version": "v260330",
    "description": "可以通过 HTTP 请求来调用下载器的 API，现在提供了以下几类接口：\n\n- 视频号相关\n- 公众号相关\n- 下载器相关\n- 下载任务相关\n\n**注意**：接口文档已拆分为多个子文件：\n- `channels.json` - 视频号接口\n- `mp.json` - 公众号接口\n- `task.json` - 下载器接口\n- `download_task.json` - 下载任务接口（/api/download_task/*）\n- `filehelper.json` - 文件传输助手接口\n- `media.json` - 媒体播放接口"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:2022",
      "description": "本地服务器"
    }
  ],
...
  "paths": {
    "/api/channels/contact/search": {
      "get": {
        "tags": ["视频号"],
        "summary": "搜索视频号账号",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "搜索关键词"
          },
          {
            "name": "next_marker",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "分页标记，用于获取下一页数据"
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {
                        "errCode": 0,
                        "errMsg": "ok",
                        "data": {
                          "BaseResponse": {
                            "Ret": 0,
                            "ErrMsg": {
                              "String": ""
                            }
                          },
                          "infoList": [
                            {
                              "contact": {
                                "username": "v2_060000231003b20faec8c4e48e1dc3ddcd03ec3cb077bb11b3c6c9a42ee3cea8073a64e6e2bd@finder",
                                "nickname": "我不叫龙虾.",
                                "headUrl": "https://wx.qlogo.cn/finderhead/ver_1/33icjVbn3Wgiccaq0iamxIenIO3ZucxmE4TtBu1WrSoUm1lnhrv2wmL8S7wxfNkPqrjSsmxcHx7aZFoWYNlNlsoYZFPVX3ichOv49HnS8EHIWY5MQnppdhncQkE5flrIe38F/132",
                                "signature": "我不叫龙虾🦞",
                                "coverImgUrl": ""
                              },
                              "highlightNickname": "我不叫<em class=\"highlight\">龙虾</em>.",
                              "highlightSignature": "我不叫<em class=\"highlight\">龙虾</em>🦞",
                              "friendFollowCount": 0,
                              "highlightProfession": "国风自媒体",
                              "reqIndex": 0
                            },
                            {
                              "contact": {
                                "username": "v2_060000231003b20faec8c4e08c10cad4c707ee37b077c5f14e385655f6f7ed18eaea3cbfe7b7@finder",
                                "nickname": "辣个龙虾.海鲜老板娘baby洋",
                                "headUrl": "https://wx.qlogo.cn/finderhead/ver_1/SRnmI6Zgrj53fmHlMEl3gX71hm867zy2GB0iaDxrF8bT3anoKd32XbN4OkQqFNn6CMS0BUkgJcQNXfWiaFIUAA5DPmE9M7HXLaZDxZVE913tFrp1ibLU0xsDTicuxIY3ibiaOR/132",
                                "signature": "合肥包河区宁国路大宁缤购商业广场停车场入口。辣个龙虾海鲜，订餐电话199 6618 2806",
                                "coverImgUrl": ""
                              },
                              "highlightNickname": "辣个<em class=\"highlight\">龙虾</em>.海鲜老板娘baby洋",
                              "highlightSignature": "合肥包河区宁国路大宁缤购商业广场停车场入口。辣个<em class=\"highlight\">龙虾</em>海鲜，订餐电话199 6618 2806",
                              "friendFollowCount": 0,
                              "highlightProfession": "",
                              "reqIndex": 1
                            }
                          ]
                        }
                      }
                    }
                  },
                  "nosocket": {
                    "summary": "业务异常",
                    "value": {
                      "code": 400,
                      "msg": "请先初始化客户端 socket 连接",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/contact/feed/list": {
      "get": {
        "tags": ["视频号"],
        "summary": "获取视频号指定账号的视频列表",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "视频号 ID，例如 `v2_xxx@finder`"
          },
          {
            "name": "next_marker",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "分页标记，用于获取下一页数据"
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {
                        "errCode": 0,
                        "errMsg": "ok",
                        "data": {
                          "BaseResponse": {
                            "Ret": 0,
                            "ErrMsg": {
                              "String": ""
                            }
                          },
                          "object": [
                            {
                              "id": "14545102784038246591",
                              "contact": {
                                "username": "v2_060000231003b20faec8c4e48e1dc3ddcd03ec3cb077bb11b3c6c9a42ee3cea8073a64e6e2bd@finder",
                                "nickname": "我不叫龙虾.",
                                "headUrl": "https://wx.qlogo.cn/finderhead/ver_1/33icjVbn3Wgiccaq0iamxIenIO3ZucxmE4TtBu1WrSoUm1lnhrv2wmL8S7wxfNkPqrjSsmxcHx7aZFoWYNlNlsoYZFPVX3ichOv49HnS8EHIWY5MQnppdhncQkE5flrIe38F/132",
                                "signature": "我不叫龙虾🦞",
                                "coverImgUrl": ""
                              },
                              "objectDesc": {
                                "description": "不是我都说要嫁给你啦 你打我干嘛˃̣̣̥᷄⌓˂̣̣̥᷅",
                                "media": [
                                  {
                                    "url": "https://finder.video.qq.com/251/20302/stodownload?encfilekey=Cvvj5Ix3eewK0tHtibORqcsqchXNh0Gf3YiaX8QrZIfDfbbDuNxCaRrmC3HnECprk1pCZheGG4QmWjUCo7ibnm6KqoRAhibVpLqYY3ibiaUtyqhOZicu5gxbNT6lp9yrfM5DyX0&hy=SZ&idx=1&m=&uzid=2",
                                    "mediaType": 4,
                                    "videoPlayLen": 7,
                                    "width": 1080,
                                    "height": 1920,
                                    "fileSize": 11557355,
                                    "spec": [
                                      {
                                        "fileFormat": "xWT111",
                                        "width": 720,
                                        "height": 1280,
                                        "durationMs": 7233
                                      },
                                      {
                                        "fileFormat": "xWT112",
                                        "width": 576,
                                        "height": 1024,
                                        "durationMs": 7233
                                      },
                                      {
                                        "fileFormat": "xWT156",
                                        "width": 720,
                                        "height": 1280,
                                        "durationMs": 7233
                                      },
                                      {
                                        "fileFormat": "xWT113",
                                        "width": 576,
                                        "height": 1024,
                                        "durationMs": 7233
                                      },
                                      {
                                        "fileFormat": "xWT157",
                                        "width": 576,
                                        "height": 1024,
                                        "durationMs": 7233
                                      },
                                      {
                                        "fileFormat": "xWT158",
                                        "width": 576,
                                        "height": 1024,
                                        "durationMs": 7233
                                      }
                                    ],
                                    "coverUrl": "http://wxapp.tc.qq.com/251/20304/stodownload?encfilekey=oibeqyX228riaCwo9STVsGLIBn9G5YG8ZnffcsCiaN8QExibEBAMBrvLTnghut9ib9x6QmLCpPMl6UiaBhVwqlNmwmrDSRFyYp4ibwLWgbEepMk4Ficu4OG5DEhQtJ3wkUt3eCAJ5rbfibz901T4&token=x5Y29zUxcibDHxWfF8R3ao2MN4d8PMDvFN3hfeibV9eVnJ0LMRiana1aLYjJn0L5ibArHq9Tiawf2oW70ibEgD1OnUH94zuRLtn8ETLRBePNMz2RvXKfSMgnVsxTGUUgCBxMB1JKES9txma7xm53spLoUHgguib3g9z7H4v&idx=1&hy=SZ&m=5140e64d871f1f103a4c3a70925ae551&uzid=2&picformat=200",
                                    "decodeKey": "930220499",
                                    "urlToken": "&token=x5Y29zUxcibAq37UB3SAO7X1q6BG5zDyLxbI5IB5y0iacyPaShR0NUgpbzP0FSF7nXO2qACJa3dTcicAn8qD5z2MxKdgQoMYOVfGNasxYNxYEPoEUlz9SfgqficN1wE3WI9ntibLp5DgxcNX2B8U5fjXZnKQIRSLxicb7XU5rO2OYmicVA&basedata=CAESABoGeFdUMTExGgZ4V1QxMTIaBnhXVDE1NhoGeFdUMTEzGgZ4V1QxNTcaBnhXVDE1OCIAKgcI4DMQABgC&sign=S0-7BBO9mVW8rfuj3PSfa-2UUvt3UN-tOSZX51LRlpF77cn9CwBeJuUm25Ou4Ij8hTZ_bCjLsqwGu-FwusRrsQ&web=1&extg=10f0000&svrbypass=AAuL%2FQsFAAABAAAAAACTC%2BKXJoVQi%2FN%2B355jaRAAAADnaHZTnGbFfAj9RgZXfw6V32imHd%2F5K9TlTPEvT7Z5OfCO16AgITCPsyqIFL7GmUVzfZzD1wYus10%3D&svrnonce=1768136415"
                                  }
                                ],
                                "mediaType": 4
                              },
                              "objectNonceId": "17453311911488030792_0_0_2_33_5217681364151768_78fbb070-eeed-11f0-9e01-614e44bc5b79",
                              "source_url": "",
                              "createtime": 1733911369
                            }
                          ]
                        }
                      }
                    }
                  },
                  "nosocket": {
                    "summary": "业务异常",
                    "value": {
                      "code": 400,
                      "msg": "请先初始化客户端 socket 连接",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/feed/profile": {
      "get": {
        "tags": ["视频号"],
        "summary": "获取视频号指定视频详情",
        "description": "**注意：必须提供 `oid`+`nid` 或 `url` 或 `eid` 其中一种组合。**",
        "parameters": [
          {
            "name": "oid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "视频唯一 ID (配合 nid 使用)"
          },
          {
            "name": "nid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "视频唯一 ID (配合 oid 使用)"
          },
          {
            "name": "url",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "视频号页面 URL，https://channels.weixin.qq.com/web/pages/feed?oid=zagCB5LjCrE&nid=d3pMFaDgxy4\n\n传入的地址需要进行编码。\n\n"
          },
          {
            "name": "eid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "企微打开视频号页面时 URL 会携带的参数\n\n"
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {
                        "errCode": 0,
                        "errMsg": "ok",
                        "data": {
                          "BaseResponse": {
                            "Ret": 0,
                            "ErrMsg": {
                              "String": ""
                            }
                          },
                          "commentInfo": [],
                          "object": {
                            "id": "14819096805414996657",
                            "contact": {
                              "username": "v2_060000231003b20faec8cae08f19c0d1c802ee33b07778a2a650b6d670ef63af3f3ac1c51124@finder",
                              "nickname": "Girls-Talk",
                              "headUrl": "https://wx.qlogo.cn/finderhead/ver_1/JrkaKTpRhjYlpkDnIicnbhicmYPxsX769yKAMexpa0icN3vBq8eDh2cNl1gQtpOsI4hMz1URF9e8icMNeD3AJR080J5nURCqSedFmic4z5PMLGCTSfU6KibpwM6wTSw549pcyC/132",
                              "signature": "“女性，她们有思想，有灵魂”",
                              "coverImgUrl": ""
                            },
                            "objectDesc": {
                              "description": "“被误解是表达者的宿命，不要活在别人的语言里”#陈都灵 #女性力量 #女性成长 #女性智慧",
                              "media": [
                                {
                                  "url": "https://finder.video.qq.com/251/20302/stodownload?encfilekey=Cvvj5Ix3eewK0tHtibORqcsqchXNh0Gf3sJcaYqC2rQDpwzgf4OvFRdRpHRb1MzrzSdfFPCgFTfk8mrhzsLKe2B1JIDYnE3gQY3qRoGoRjic3p3In1NWtW5hpw1HMG6OVu&bizid=1023&dotrans=0&hy=SH&idx=1&m=&uzid=7a166",
                                  "mediaType": 4,
                                  "videoPlayLen": 41,
                                  "width": 1080,
                                  "height": 1440,
                                  "fileSize": 50833261,
                                  "spec": [
                                    {
                                      "fileFormat": "xWT111",
                                      "width": 720,
                                      "height": 960,
                                      "durationMs": 41901
                                    },
                                    {
                                      "fileFormat": "xWT112",
                                      "width": 720,
                                      "height": 960,
                                      "durationMs": 41901
                                    },
                                    {
                                      "fileFormat": "xWT113",
                                      "width": 576,
                                      "height": 768,
                                      "durationMs": 41917
                                    },
                                    {
                                      "fileFormat": "xWT126",
                                      "width": 720,
                                      "height": 960,
                                      "durationMs": 41901
                                    },
                                    {
                                      "fileFormat": "xWT127",
                                      "width": 720,
                                      "height": 960,
                                      "durationMs": 41901
                                    },
                                    {
                                      "fileFormat": "xWT128",
                                      "width": 576,
                                      "height": 768,
                                      "durationMs": 41917
                                    }
                                  ],
                                  "coverUrl": "http://wxapp.tc.qq.com/251/20304/stodownload?encfilekey=rjD5jyTuFrIpZ2ibE8T7YmwgiahniaXswqzGOtP29kPHJ59RtsyZf6v0824RAZdtLQynv4NqUjgD9I8BWbZqqBYDPhibMiayIvppYOpjUGsia85ias3YfIeYUcJeQ&token=x5Y29zUxcibDHxWfF8R3aox1Dqic3uCibRBu4icgTRABiaiaUicKibH4FDwwdebfA9asyAT9balYg6YFFicSLTubJSibAyUOgmGPDVCSwcTR5H1JicuJkYGLRFwut8yxwwCFwkuwMSK0R5w0WibxSYC2DCOJBxL56icn9DNqYDEZQ&idx=1&hy=SH&m=&scene=2&uzid=1&picformat=200",
                                  "decodeKey": "394253653",
                                  "urlToken": "&token=x5Y29zUxcibAq37UB3SAO7dZXR9anXy7Wgxcicu3bLjSsZKSTvUnmxjamgZU7ag2npbeZoVUALC0KhiaicdSfdEtWImibjOONu0CawwUCULAvnIdt9Jh1ZExd20azibwqRUzjdAL15SIdsupicNt78C1eErfO04vcEErXvLnQ6RyLyaAPg&basedata=CAESABoDeFYwIgAqBwiLKRAAGAI&sign=ttbnb0qaI_HwA0hbluHgg3lBd4WrdvcMXnfV7UzhS8eIQYJrWc2fCurJNNwqDi28J4ykLreRMQ1WEBeT9KtmTw&ctsc=146&web=1&extg=10f0000&svrbypass=AAuL%2FQsFAAABAAAAAACZ3xK%2FuPwmyeBpAppjaRAAAADnaHZTnGbFfAj9RgZXfw6V996Fck4oFO1qxgm3OHC7cAgwlUePFkt0tDtvlNPSpCMbnl54fJ6Crxs%3D&svrnonce=1768135170"
                                }
                              ],
                              "mediaType": 4
                            },
                            "objectNonceId": "8609277293449234222_0_146_0_0",
                            "source_url": "",
                            "createtime": 1766574001
                          },
                          "commentCount": 193,
                          "nextCheckObjectStatus": 1,
                          "barrageCommentInfo": [],
                          "refObjectList": [],
                          "preloadInfo": {
                            "preloadStrategyId": 17840794003360850000,
                            "globalInfo": {
                              "prevCount": 0,
                              "nextCount": 5,
                              "maxBitRate": 150,
                              "preloadFileSizePercent": 0.25,
                              "preloadFileMinBytes": 0,
                              "preloadMaxConcurrentCount": 1,
                              "megavideoMaxBitRate": 0,
                              "megavideoPrevCount": 1,
                              "megavideoNextCount": 2,
                              "minBufferLength": 10,
                              "maxBufferLength": 10,
                              "minCurrentFeedBufferLength": 5,
                              "minRefreshInterval": 86400,
                              "preloadFileTimeMs": 30000,
                              "localserverBufferSize": 524288,
                              "localserverSendSocketBufferSize": 131072,
                              "ffmpegTcpRecvSocketBufferSize": 131072,
                              "lastFeedInfoCount": 12,
                              "nextFeedInfoCount": 1,
                              "allowToUseNewSpecMaxPercent": -1,
                              "isBackendControlFindPagePreload": false,
                              "isFindPagePreload": false,
                              "findPagePreloadSecond": 5,
                              "wcplayerSchedulerType": 2,
                              "minStartPlayBufferLengthMs": 4000,
                              "findPageEnableCellularPreload": false,
                              "holdoutExptParamSet": {
                                "holdoutItems": [
                                  {
                                    "key": "holdout_expt_2",
                                    "value": "default_test_2",
                                    "inPriority": false
                                  },
                                  {
                                    "key": "holdout_expt_1",
                                    "value": "default_test",
                                    "inPriority": true
                                  }
                                ],
                                "hitHoldoutExpt": 0
                              },
                              "flowViewPreload": true,
                              "bufferingHealthLengthL0": 5,
                              "bufferingHealthLengthL1": 10,
                              "bufferingHealthLengthL2": 15,
                              "loadWithPlaying": [],
                              "loadWithPlayWait": [],
                              "findPagePreloadFeedCount": 1
                            },
                            "objectInfo": [
                              {
                                "feedId": 14819096805414996000,
                                "interestFactor": 1,
                                "loadWithPlaying": []
                              }
                            ],
                            "megavideoInfo": []
                          },
                          "traceBuffer": "CLGVjJf5wIDUzQE=",
                          "liveAliasInfo": [],
                          "descCommentInfo": [],
                          "userTags": [],
                          "reportBypass": "{\"client_region\":{\"countryid\":\"cn\",\"ipaddr\":\"183.129.150.86\"}}"
                        },
                        "payload": {
                          "needObject": 1,
                          "lastBuffer": "",
                          "scene": 146,
                          "direction": 2,
                          "identityScene": 2,
                          "pullScene": 6,
                          "objectid": "14819096805414996657",
                          "objectNonceId": "8609277293449234222",
                          "encrypted_objectid": ""
                        }
                      }
                    }
                  },
                  "nosocket": {
                    "summary": "业务异常",
                    "value": {
                      "code": 400,
                      "msg": "请先初始化客户端 socket 连接",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/channels/interactioned/list": {
      "get": {
        "tags": ["视频号"],
        "summary": "获取我的赞和收藏列表",
        "parameters": [
          {
            "name": "next_marker",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "分页标记"
          },
          {
            "name": "flag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "enum": [1, 2, 4, 7]
            },
            "description": "赞、喜欢、收藏标记。1: 喜欢, 2: 赞, 4: 收藏, 7: 全部"
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {
                        "errCode": 0,
                        "errMsg": "ok",
                        "data": {
                           "object": []
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rss/channels": {
      "get": {
        "tags": ["视频号"],
        "summary": "RSS 订阅",
        "description": "获取指定视频号的 RSS Feed。",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "视频号 ID"
          },
          {
            "name": "next_marker",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "分页标记"
          }
        ],
        "responses": {
          "200": {
            "description": "RSS XML",
            "content": {
              "application/xml": {
                "example": "<feed></feed>"
              }
            }
          }
        }
      }
    },
    "/play": {
      "get": {
        "tags": ["视频号"],
        "summary": "视频代理播放",
        "description": "url 是视频地址，非「视频号页面地址」，同样需要对地址进行编码",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "目标视频地址"
          },
          {
            "name": "key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "解密密钥"
          }
        ],
        "responses": {
          "200": {
            "description": "视频流"
          }
        }
      }
    },
...
    "/api/mp/delete": {
      "post": {
        "tags": ["公众号"],
        "summary": "删除公众号",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "鉴权 Token (开启鉴权时必填)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "biz": {
                    "type": "string",
                    "description": "公众号唯一 ID"
                  }
                },
                "required": ["biz"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {}
                    }
                  },
                  "fail": {
                    "summary": "失败",
                    "value": {
                      "code": 400,
                      "msg": "参数错误",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/download_task/play": {
      "get": {
        "tags": ["下载任务"],
        "summary": "在线播放已下载文件（兼容 wx_channels_api）",
        "parameters": [
          { "name": "id", "in": "query", "required": true, "schema": { "type": "integer" }, "description": "download_task 表主键 id" }
        ],
        "responses": {
          "200": {
            "description": "文件流（或业务错误时返回 JSON）"
          }
        }
      }
    },
...
    "/channels/task/status": {
      "get": {
        "tags": ["平台"],
        "summary": "查询任务状态（兼容 wx_channels_api）",
        "parameters": [{ "name": "task_id", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "状态信息" } }
      }
    },
...
  }
}