# internal/api/routes.go
package api

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

func (c *APIClient) SetupRoutes() {
	// favicon
	c.engine.GET("/favicon.ico", c.handleFavicon)
	// 只在本地有的接口
	if !c.cfg.RemoteServerMode {
		// 视频号接口
		c.engine.GET("/api/channels/contact/search", c.handleSearchChannelsContact)
		c.engine.GET("/api/channels/contact/feed/list", c.handleFetchFeedListOfContact)
		c.engine.GET("/api/channels/feed/profile", c.handleFetchFeedProfile)
		c.engine.GET("/api/channels/live/replay/list", c.handleFetchLiveReplayList)
		c.engine.GET("/api/channels/interactioned/list", c.handleFetchInteractionedFeedList)
		c.engine.GET("/rss/channels", c.handleFetchFeedListOfContactRSS)
		// 公众号接口
		c.engine.GET("/ws/mp", c.official.HandleWebsocket)
		c.engine.GET("/ws/manage", c.official.HandleManageWebsocket)
		c.engine.POST("/api/mp/refresh_with_frontend", c.official.HandleRefreshOfficialAccountWithFrontend)
		c.engine.GET("/api/mp/ws_pool", c.official.HandleFetchOfficialAccountClients)
		// 文件传输助手接口
		c.engine.GET("/filehelper", c.filehelper.HandlePage)
		c.engine.GET("/api/filehelper/qrcode", c.filehelper.HandleGetQRCode)
		c.engine.GET("/api/filehelper/login/wait", c.filehelper.HandleWaitLogin)
		c.engine.GET("/api/filehelper/status", c.filehelper.HandleGetStatus)
		c.engine.GET("/api/filehelper/synccheck", c.filehelper.HandleSyncCheck)
		c.engine.GET("/api/filehelper/sync", c.filehelper.HandleSyncMessages)
		c.engine.GET("/api/filehelper/messages", c.filehelper.HandleGetMessages)
		c.engine.POST("/api/filehelper/send", c.filehelper.HandleSendMessage)
		c.engine.POST("/api/filehelper/logout", c.filehelper.HandleLogout)
		c.engine.POST("/api/filehelper/parse_finder_feed", c.filehelper.HandleParseFinderFeed)
		// 文件操作
		c.engine.POST("/api/show_file", c.handleHighlightFileInFolder)
		c.engine.POST("/api/open_download_dir", c.handleOpenDownloadDir)
	}
	// 下载任务接口
	c.engine.GET("/ws/downloader", c.downloader_ws.HandleDownloaderWebsocket)
	c.engine.GET("/ws/channels", c.channels.HandleChannelsWebsocket)
	c.engine.POST("/api/browse_history/create", c.handleCreateBrowseHistory)
	c.engine.POST("/api/browse_history/list", c.handleFetchBrowseHistoryList)
	c.engine.GET("/api/task/list", c.handleFetchTaskList)
	c.engine.GET("/api/task/profile", c.handleFetchTaskProfile)
	c.engine.POST("/api/task/create", c.handleCreateFeedDownloadTask)
	c.engine.POST("/api/task/create_batch", c.handleBatchCreateTask)
	c.engine.POST("/api/task/create_channels", c.handleCreateChannelsTask)
	// c.engine.POST("/api/task/create_live", c.handleCreateLiveTask)
	c.engine.POST("/api/task/create2", c.handleCreateDownloadTask)
	c.engine.POST("/api/task/start", c.handleStartTask)
	c.engine.POST("/api/task/pause", c.handlePauseTask)
	c.engine.POST("/api/task/resume", c.handleResumeTask)
	c.engine.POST("/api/task/delete", c.handleDeleteTask)
	c.engine.POST("/api/task/clear", c.handleClearTasks)
	c.engine.GET("/api/file", c.handleFetchFile)

	c.engine.POST("/api/download_task/list", c.handleCompatDownloadTaskList)
	c.engine.POST("/api/download_task/start", c.handleCompatDownloadTaskStart)
	c.engine.POST("/api/download_task/profile", c.handleCompatDownloadTaskProfile)
	c.engine.POST("/api/download_task/create", c.handleCompatDownloadTaskCreate)
	c.engine.POST("/api/download_task/batch_create", c.handleCompatDownloadTaskBatchCreate)
	c.engine.POST("/api/download_task/delete", c.handleCompatDownloadTaskDelete)
	c.engine.POST("/api/download_task/retry", c.handleCompatDownloadTaskRetry)
	c.engine.POST("/api/download_task/pause_all", c.handleCompatDownloadTaskPauseAll)
	c.engine.POST("/api/download_task/start_all", c.handleCompatDownloadTaskStartAll)
	c.engine.POST("/api/download_task/highlight_file", c.handleCompatDownloadTaskHighlightFile)
	c.engine.GET("/api/download_task/play", c.handleCompatDownloadTaskPlay)

	c.engine.POST("/browse_history/create", c.handleCreateBrowseHistory)
	c.engine.POST("/browse_history/list", c.handleFetchBrowseHistoryList)

	c.engine.GET("/influencers", c.handleCompatInfluencerList)
	c.engine.GET("/influencers/:id", c.handleCompatInfluencerGet)
	c.engine.POST("/influencers", c.handleCompatInfluencerCreate)
	c.engine.PUT("/influencers/:id", c.handleCompatInfluencerUpdate)

	c.engine.POST("/account/list", c.handleCompatAccountList)
	c.engine.POST("/account/synchronize", c.handleCompatAccountSynchronize)

	c.engine.POST("/video/list", c.handleCompatVideoList)

	c.engine.GET("/channels/search/author", c.handleCompatChannelsSearchAuthor)
	c.engine.GET("/channels/author/videos", c.handleCompatChannelsAuthorVideos)
	c.engine.GET("/channels/media/profile", c.handleCompatChannelsMediaProfile)
	c.engine.GET("/channels/task/status", c.handleCompatChannelsTaskStatus)
	c.engine.GET("/channels/task/start", c.handleCompatChannelsTaskStart)
	// 文件操作
	c.engine.GET("/play", c.handlePlay)
	c.engine.GET("/file", c.handleStreamVideo)
	c.engine.GET("/preview", c.handlePreviewFile)
	// 公众号接口 远端和本地都有的接口
	c.engine.GET("/api/mp/list", c.official.HandleFetchList)
	c.engine.GET("/api/mp/msg/list", c.official.HandleFetchMsgList)
	c.engine.GET("/api/mp/article/list", c.official.HandleFetchArticleList)
	c.engine.POST("/api/mp/delete", c.official.HandleDelete)
	c.engine.POST("/api/mp/refresh", c.official.HandleRefreshEvent)
	c.engine.GET("/rss/mp", c.official.HandleOfficialAccountRSS)
	c.engine.GET("/mp/proxy", c.official.HandleOfficialAccountProxy)
	c.engine.GET("/mp/home", c.official.HandleOfficialAccountManagerHome)
	// 其他
	c.engine.GET("/api/status", c.handleStatus)
	// c.engine.GET("/api/test", c.handleTest)

	c.engine.NoRoute(func(ctx *gin.Context) {
		ctx.Header("Content-Type", "text/html; charset=utf-8")
		ctx.String(http.StatusNotFound, "<!doctype html><html lang=\"zh-CN\"><head><meta charset=\"utf-8\"><meta name=\"viewport\" content=\"width=device-width, initial-scale=1\"><title>404 Not Found</title><style>body{margin:0;font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Helvetica,Arial,sans-serif;background:#0b0c0f;color:#e6e6e6;display:flex;align-items:center;justify-content:center;height:100vh}.box{max-width:560px;padding:24px 28px;border-radius:12px;background:#14171f;box-shadow:0 8px 24px rgba(0,0,0,.3)}h1{margin:0 0 8px;font-size:24px}p{margin:0;color:#b0b0b0}a{color:#8ab4f8;text-decoration:none}a:hover{text-decoration:underline}</style></head><body><div class=\"box\"><h1>404 未找到页面</h1><p>请求的路径不存在。返回 <a href=\"/\">首页</a></p></div></body></html>")
	})
}

func (c *APIClient) handleFavicon(ctx *gin.Context) {
	ctx.Header("Content-Type", "image/png")
	ctx.Header("Cache-Control", "public, max-age=86400")
	ctx.File("winres/icon.png")
}

func (c *APIClient) handleStatus(ctx *gin.Context) {
	channels_data := gin.H{
		"available": false,
	}
	err := c.channels.Validate()
	if err != nil {
		channels_data["available"] = false
	}
	data := gin.H{
		"version":  c.cfg.Version,
		"channels": channels_data,
	}
	ctx.JSON(200, gin.H{
		"code": 0,
		"msg":  "ok",
		"data": data,
	})
}